visitor_app/requirements.txt
streamlit
mysql-connector-python
boto3
bcrypt
qrcode
Pillow
streamlit-drawable-canvas
python-dotenv
